Adding Homescreen and PieChart
Because:
- To create a basis for the app to expand on

Creating a home screen as the basis for the project. The view contains a pie chart and a text area
where users can see their meeting times and the progress regarding the collection of VP hours. The
navigation items in form of 2 buttons and a pie chart were added and click listeners implemented.
Additionally there is a load function for the slices of the pie chart.

https://www.geeksforgeeks.org/how-to-add-a-pie-chart-into-an-android-application/
Co-Author: Niklas