Adding the feature to create studies
Because:
- Enabling user to create their own studies
- Give guidance for relevant information

The creation of studies within the app provides information of what should be included in the
description of a study so all participants have enough information to partake(#17). This also
involves the opportunity to add VP-hours, which are offered for taking part (#19). If mandatory
information are missing the system notifies the user and allows to provide the missing information.
If there is missing information, which is not mandatory the user will get a notice to draw his
attention to it and allows him to fix if needed(#36). The creation also provides a tool to add
dates in which the study will take place (#27).