Creating the view for the personal account of the user
Because:
- the users could not see the completed studies (issue #2)
- the users could not see all the studies in which they applied for participation and how many VP hours they will have after that (issue #29)
- the users could not see the when they have an appointment

Adding a pieChart and a expandable list to the activity. The pieChart shows the VP hours of the user, split into completed, participated and planned. Below is the expandable list with all the studies under each of the three categories. Additionally there is a forth category with the all the studies the user created. The list items show the title of the study, how many VP hours they bring and when the appointment is set (issue #3)
Clicking on a Study will lead to the detail view of the study.

close #2, close #3, close #29