Added the navigation drawer to the home activity (issue #14)
Because:
- The users should use this drawer to navigate through the app

Added the navigation drawer and adjusted the app bar in the home activity_home.xml to be able to trigger the drawer.
Added icons to visualize the items in the drawer.
The drawer can be accessed by swiping right or pressing the burger menu button in the app bar.