The user can now only select a date if it isn't already selected
Because:
- It was possible to overwrite somebody else's date

If the user loads a detailed view of a study all available dates are displayed.
When in the meantime another user selects a date of this study, this date is still displayed in the list of the first user.
Now, when a date is already selected, it isn't possible to overwrite it. The user is also notified with a snackBar about the status of a date's selection or deselection.