filterlists completed
because:
- displaying issues when colorfilter was used
- displaying issues when using colorfilter
- displaying issues when using Name, date or vps filter

Due to a circled call of methods the sorted lists were reset to default whenever a filter was activated.
When activating a color filter, only visible items were filtered. Changed listviewitem call from listview.getChildAt() to adapter.getObjects.get() in order to get all items and not only the visible ones.