bug fix
because:
- sign out back button undid the sign out
- findStudy => studydetails back navigaion opened navigation drawer
- back button opened drawer

When backtracking from StudyDetailsFragment to findStudyFragment the navigation drawer was opened. The back button opened the navigation drawer in all main Fragments and did not close it on repress.

When using the sign out option in the navigation drawer the user could sign in by using the phones back button